Reorganized code, moved model loading to new folder
